Ran flake8 to carry out linting

+++ crud_app/models/crudmodel.py
@@ -4,12 +4,15 @@ This is just a model sample. It depends on which database you want to use but
 basically make sure this file only contains methods and classes that are
 related to this model.
 """
-from oto import response
+
 from sqlalchemy import Column
 from sqlalchemy import Integer
 from sqlalchemy import String
 
+from oto import response
+
 from crud_app.connectors import mysql
+
 
 
 class Node(mysql.BaseModel):
@@ -35,11 +38,11 @@ class Node(mysql.BaseModel):
         return {
             'id': self.test_id,
             'name': self.name,
-            'surname':self.surname
+            'surname': self.surname
         }
 
 
-def get_by_id(model_id):
+# def get_by_id(model_id):
     """Get a model by its id.
 
     Note:
@@ -54,8 +57,9 @@ def get_by_id(model_id):
     Returns:
         response.Response: the data of the model.
     """
-    model = request.get('ows-microservice', '/resource')
-    return response.Response(message=model.json(), status=model.status_code)
+    # model = request.get('ows-microservice', '/resource')
+    # return response.Response(message=model.json(), status=model.status_code)
+
 
 def get_nodes_all():
     """Get all nodes
@@ -69,10 +73,11 @@ def get_nodes_all():
             result_set = session.query(Node).all()
 
             if not result_set:
-                return response.Response('test data not available.')
+                return response.Response('No nodes found')
 
             total_records = [r.to_dict() for r in result_set]
             return response.Response(message=total_records)
+
 
 def get_node_for(nodeid):
     """Get node for passed node id
@@ -86,13 +91,12 @@ def get_node_for(nodeid):
     with mysql.db_session() as session:
         result_set = session.query(Node).get(nodeid)
         if result_set:
-            node_name = result_set.name
-            node_sname = result_set.surname
             nodejson = result_set.to_dict()
         else:
-            nodejson = {"message":"No Node exists for given nodeid"}
+            nodejson = {'message': 'No Node exists for given nodeid'}
 
     return response.Response(message=nodejson)
+
 
 def delete_node(nodeid):
     """Delete node for passed node id
@@ -105,13 +109,17 @@ def delete_node(nodeid):
     """
     with mysql.db_session() as session:
         result_set = session.query(Node).get(nodeid)
-        node_name = result_set.name
-        nodejson = result_set.to_dict()
-        session.delete(result_set)
+        if result_set:
+            nodejson = result_set.to_dict()
+            session.delete(result_set)
+            response_message = ('Deleted node {}').format(nodejson)
+        else:
+            response_message = 'No node found'
 
-    return response.Response(message=("Deleted node {}").format(nodejson))
+    return response.Response(message=response_message)
 
-def update_node(nodeid,paramdict):
+
+def update_node(nodeid, paramdict):
     """Update node for passed node id
 
     Args:
@@ -122,16 +130,19 @@ def update_node(nodeid,paramdict):
     """
     with mysql.db_session() as session:
         result_set = session.query(Node).get(nodeid)
-        # keyvalue_tuples = param_imdict[0]
-        # paramdict = keyvalue_tuples # Convert immutable paramter tuples into a key value dictionary
-        # paramdict = {t[0]:t[1] for t in keyvalue_tuples} # Convert immutable paramter tuples into a key value dictionary
-        node_name = result_set.name
-        nodejson = result_set.to_dict()
-        for colname,colval in paramdict.items():
-            setattr(result_set,colname,colval)
-        session.merge(result_set)
+        if result_set:
+            nodejson = result_set.to_dict()
+            for colname, colval in paramdict.items():
+                setattr(result_set, colname, colval)
+            session.merge(result_set)
+            response_message = message=('Updated node {} with \
+                                {}').format(nodejson, paramdict)
+        else:
+            response_message = 'No node found'
 
-    return response.Response(message=("Updated node {} with {}").format(nodejson,paramdict))
+
+    return response.Response(response_message)
+
 
 def create_node(paramdict):
     """Update node for passed node id
@@ -139,19 +150,15 @@ def create_node(paramdict):
     Args:
             paramdict: Dictionary column value pairs to be added as a row
     Returns:
-        response.Response: the node 
+        response.Response: the node
     """
-
-    # valid_paramdict = [k:v for k,v in paramdict if k in ]
     with mysql.db_session() as session:
         try:
             new_node = Node(**paramdict)
             session.add(new_node)
-            response_message = "Successfully added"
+            response_message = 'Successfully added'
         except:
-            response_message = "Please add correct key value pairs"
+            response_message = 'Please add correct key value pairs'
     session.close()
 
     return response.Response(response_message)
-
-
